Extract button components to their own files.
