+++ client/src/main/kotlin/io/spine/examples/chatspn/desktop/component/TextButton.kt
@@ -26,8 +26,6 @@
 
 package io.spine.examples.chatspn.desktop.component
 
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.size
@@ -38,7 +36,6 @@ import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.vector.ImageVector
@@ -83,38 +80,8 @@ public fun TextButton(
         }
         Text(
             text,
+            color = contentColor,
             style = MaterialTheme.typography.bodyLarge
         )
     }
-}
-
-/**
- * Displays the icon button.
- *
- * @param icon icon to display
- * @param onClick callback that will be triggered when the icon clicked
- * @param color icon color
- * @param modifierExtender extension for the modifier
- */
-@Composable
-public fun IconButton(
-    icon: ImageVector,
-    color: Color = MaterialTheme.colorScheme.primary,
-    modifierExtender: Modifier.() -> Modifier = { this },
-    onClick: () -> Unit,
-) {
-    val interactionSource = remember { MutableInteractionSource() }
-    Icon(
-        icon,
-        "Input icon",
-        Modifier
-            .pointerHoverIcon(PointerIcon(Cursor.getPredefinedCursor(Cursor.HAND_CURSOR)))
-            .clickable(
-                interactionSource = interactionSource,
-                indication = null,
-                onClick = onClick
-            )
-            .modifierExtender(),
-        color
-    )
 }
